requirements.txt
Flask==2.2.2
python-dotenv==1.0.0
mongoengine==0.27.0
Flask-Admin==1.6.1
Flask-MongoEngine==1.0.0
Flask-Cors==4.0.0
boto3==1.34.40
gunicorn==21.2.0
Werkzeug==2.2.2
pymongo<4
WTForms==2.3.3